- constants.h now defines DMTCP_FAIL_RC.  Whenever dmtcp_checkpoint or    dmtcp_restart fails, it should return this, to distinguish from    the user program failing with its own return particular code. - minor polishing of pthread2.c, autogen.sh, jsocket.cpp, mtcp.c

+++ trunk/test/pthread2.c
@@ -43,14 +43,14 @@ static void *threadMain (void *data)
   int id = *(int*) data;
 
   while (1) {
-    printf("Worker: %d (%d) alive. numWorkers: %d\n",
-           id, syscall(SYS_gettid), numWorkers);
+    printf("Worker: %d (%ld) alive. numWorkers: %d\n",
+           id, (long)syscall(SYS_gettid), numWorkers);
     usleep(100*1000);
     pthread_mutex_lock(&mutex);
     if (numWorkers > 5) {
       numWorkers--;
       printf("Worker: %d (%ld) exiting: numWorkers: %d\n",
-             id, (long) syscall(SYS_gettid), numWorkers);
+             id, (long)syscall(SYS_gettid), numWorkers);
       pthread_mutex_unlock(&mutex);
       free(data);
       pthread_exit(NULL);
